[ADD] nubosoft: Added tax withholding format
Adding the dependency `l10n_mx_edi_tax_withholding` and the Report for
payment with Tax Withholding.

A method `_convert_amount_to_payment_currency` was also added to the
`account.payment` model to use it in the report because the amounts in
the CFDI come already converted to the Company's currency, and we need
it in the payment's currency.

Authored-By: Luis Escobar <lescobar@vauxoo.com>
Co-Authored-By: Carmen Miranda <carmen@vauxoo.com>